refactor(packages/cli): extract dev lifecycle into useDevServer hook
Move all server lifecycle state, refs, and side effects from DevScreen
into a dedicated useDevServer hook. DevScreen is now a pure render
function that wires state to UI and hotkeys to actions.

- New: packages/cli/src/hooks/use-dev-server.ts
- New types: DevPhase, DevServerState, DevServerActions, LogEntry
- DevScreen reduced from 434 to 225 lines (pure render + hotkeys)
